devices: add full support for APA/HDL by using PFS partition as metadata device, move device-specific functions to separate files; add support for syncing devices
